src/test/java/com/numericalmethod/suanshu/stats/random/univariate/poisson/Knuth1969Test.java
/*
 * Copyright (c) Numerical Method Inc.
 * http://www.numericalmethod.com/
 * 
 * THIS SOFTWARE IS LICENSED, NOT SOLD.
 * 
 * YOU MAY USE THIS SOFTWARE ONLY AS DESCRIBED IN THE LICENSE.
 * IF YOU ARE NOT AWARE OF AND/OR DO NOT AGREE TO THE TERMS OF THE LICENSE,
 * DO NOT USE THIS SOFTWARE.
 * 
 * THE SOFTWARE IS PROVIDED "AS IS", WITH NO WARRANTY WHATSOEVER,
 * EITHER EXPRESS OR IMPLIED, INCLUDING, WITHOUT LIMITATION,
 * ANY WARRANTIES OF ACCURACY, ACCESSIBILITY, COMPLETENESS,
 * FITNESS FOR A PARTICULAR PURPOSE, MERCHANTABILITY, NON-INFRINGEMENT, 
 * TITLE AND USEFULNESS.
 * 
 * IN NO EVENT AND UNDER NO LEGAL THEORY,
 * WHETHER IN ACTION, CONTRACT, NEGLIGENCE, TORT, OR OTHERWISE,
 * SHALL THE AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR
 * ANY CLAIMS, DAMAGES OR OTHER LIABILITIES,
 * ARISING AS A RESULT OF USING OR OTHER DEALINGS IN THE SOFTWARE.
 */
package com.numericalmethod.suanshu.stats.random.univariate.poisson;

import com.numericalmethod.suanshu.stats.descriptive.moment.Kurtosis;
import com.numericalmethod.suanshu.stats.descriptive.moment.Mean;
import com.numericalmethod.suanshu.stats.descriptive.moment.Skewness;
import com.numericalmethod.suanshu.stats.descriptive.moment.Variance;
import com.numericalmethod.suanshu.stats.distribution.univariate.PoissonDistribution;
import org.junit.Test;
import static org.junit.Assert.*;

/**
 *
 * @author Haksun Li
 */
public class Knuth1969Test {

    @Test
    public void test_0010() {
        final int N = 10000;
        double lambda = 1;

        Knuth1969 instance = new Knuth1969(lambda);
        instance.seed(123456789L);

        double[] sample = new double[N];
        for (int i = 0; i < N; ++i) {
            sample[i] = instance.nextDouble();
        }

        double mean = new Mean(sample).value();
        double var = new Variance(sample).value();
        double skew = new Skewness(sample).value();
        double kurtosis = new Kurtosis(sample).value();

        PoissonDistribution dist = new PoissonDistribution(lambda);
        assertEquals(dist.mean(), mean, 5e-2);
        assertEquals(dist.variance(), var, 5e-2);
        assertEquals(dist.skew(), skew, 5e-2);
        assertEquals(dist.kurtosis(), kurtosis, 5e-2);
    }
}
